Rename WdigetRendererComponent to SmartUiComponent
